% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment2.R
\docType{data}
\name{data}
\alias{data}
\title{Descriptive data title}
\format{
a \code{tbl_df} containing the following fields:
\describe{
\item{id}{}
\item{alignment}{}
\item{reliability_amount}{}
\item{start_date}{}
\item{end_date}{}
\item{duration_in_seconds}{}
\item{response_id}{}
\item{sex}{}
\item{age}{}
\item{language}{}
\item{language_other}{}
\item{country}{}
\item{school}{}
\item{justification}{}
\item{project}{}
\item{ranking}{}
\item{allocation}{}
\item{confidence}{}
\item{weighting}{}
\item{forecast}{}
\item{date_rank}{}
\item{npv_amount}{}
\item{sample}{}
\item{weighted_forecast}{}
\item{forecast_mean}{}
\item{forecast_sd}{}
\item{max_npv}{}
\item{min_npv}{}
\item{max_min_difference}{}
}
}
\source{
The data comes from________________________.
}
\description{
Detailed description of the data
}
\seealso{
\link{alignment2}
}
