% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dv_replacement.R
\name{get_dv_replacement}
\alias{get_dv_replacement}
\title{Get DV replacement for Experiment 3}
\usage{
get_dv_replacement(data_raw, dv, dv_id)
}
\arguments{
\item{data_raw}{}

\item{dv_id}{}
}
\value{

}
\description{
Get DV replacement for Experiment 3
}
\author{
Shir Dekel
}
