% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_data.R
\name{pivot_data}
\alias{pivot_data}
\title{Pivot data

Seems we have to pivot this many times because justification isn't split by project and also because of the way dv_replacement is coded for forecasts}
\usage{
pivot_data(data, dv_replacement, dv_id)
}
\arguments{
\item{data}{}

\item{dv_id}{}
}
\value{

}
\description{
Pivot data

Seems we have to pivot this many times because justification isn't split by project and also because of the way dv_replacement is coded for forecasts
}
\author{
Shir Dekel
}
