% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_age.R
\name{recode_age}
\alias{recode_age}
\title{Recode age variable

The max is 1976, so I'm assuming the participant entered his year of birth. Subsequent experiments restricted age values over 100.}
\usage{
recode_age(age)
}
\arguments{
\item{age}{}
}
\value{

}
\description{
Recode age variable

The max is 1976, so I'm assuming the participant entered his year of birth. Subsequent experiments restricted age values over 100.
}
\author{
Shir Dekel
}
