% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_alignment.R
\name{recode_alignment}
\alias{recode_alignment}
\title{Recode alignment}
\usage{
recode_alignment(alignment)
}
\arguments{
\item{alignment}{}
}
\value{

}
\description{
Recode alignment
}
\author{
Shir Dekel
}
