% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_weighting.R
\name{recode_weighting}
\alias{recode_weighting}
\title{Recode weighting}
\usage{
recode_weighting(weighting)
}
\arguments{
\item{weighting}{}
}
\value{

}
\description{
Recode weighting
}
\author{
Shir Dekel
}
