% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_qualtrics.R
\name{rename_qualtrics}
\alias{rename_qualtrics}
\title{Rename qualtrics data}
\usage{
rename_qualtrics(data)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
Rename qualtrics data
}
\author{
Shir Dekel
}
