% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_pivot_join_experiment3.R
\name{split_pivot_join_experiment3}
\alias{split_pivot_join_experiment3}
\title{Split, pivot, and join Experiment 3 data}
\usage{
split_pivot_join_experiment3(data, dv)
}
\value{

}
\description{
Split, pivot, and join Experiment 3 data
}
\author{
Shir Dekel
}
