% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment2.R
\docType{package}
\name{alignment2}
\alias{alignment2}
\alias{alignment2-package}
\title{Alignment 2 experiment}
\description{
Data and materials for the Alignment 2 experiment.
}
\details{
alignment2
A data package for the Alignment 2 experiment.

Use \code{data(package='alignment2')$results[, 3]} to see a list of available data sets in this data package
and/or DataPackageR::load_all
_datasets() to load them.
}
\seealso{
\link{data}
}
